make Selector type consistent now it always takes an object
- split up @qawolf/types
- require Event.page
