+++ packages/browser/src/find/find.ts
@@ -1,19 +1,9 @@
-import { FindOptions } from "@qawolf/types";
+import { FindOptions, Selector } from "@qawolf/types";
 import { ElementHandle, Page } from "puppeteer";
 import { findCss } from "./findCss";
-import { findHtml, HtmlSelector } from "./findHtml";
+import { findHtml } from "./findHtml";
 import { findText } from "./findText";
 import { retryExecutionError } from "../retry";
-
-// https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Selectors
-export type CssSelector = string;
-
-export type Selector =
-  | CssSelector
-  | {
-      html?: HtmlSelector;
-      text?: string;
-    };
 
 export const find = (
   page: Page,
@@ -21,8 +11,8 @@ export const find = (
   options: FindOptions
 ): Promise<ElementHandle | null> => {
   return retryExecutionError(async () => {
-    if (typeof selector === "string") {
-      return findCss(page, selector, options);
+    if (selector.css) {
+      return findCss(page, selector.css, options);
     }
 
     if (selector.html) {
