+++ packages/browser/tests/actions/select.test.ts
@@ -22,7 +22,7 @@ describe("select", () => {
     });
     expect(selectValue).toBeFalsy();
 
-    const element = await browser.find("#dropdown");
+    const element = await browser.find({ css: "#dropdown" });
     await select(element, "2");
 
     const selectValue2 = await page.evaluate(() => {
@@ -33,14 +33,14 @@ describe("select", () => {
   });
 
   it("throws error if option with value not available before timeout", async () => {
-    const element = await browser.find("#dropdown");
+    const element = await browser.find({ css: "#dropdown" });
 
     const testFn = async () => await select(element, "3", 2000);
     await expect(testFn()).rejects.toThrowError();
   });
 
   it("throws error if option with value available but disabled before timeout", async () => {
-    const element = await browser.find("#dropdown");
+    const element = await browser.find({ css: "#dropdown" });
 
     const testFn = async () => await select(element, "", 2000);
     await expect(testFn()).rejects.toThrowError();
