+++ packages/build-workflow/src/buildClickSteps.ts
@@ -1,6 +1,6 @@
 import { logger } from "@qawolf/logger";
-import { Event, KeyEvent, Step, Doc } from "@qawolf/types";
-import { isKeyEvent, isTypeEvent } from "@qawolf/web";
+import { Event, Step } from "@qawolf/types";
+import { isTypeEvent } from "@qawolf/web";
 
 export const buildClickSteps = (events: Event[]): Step[] => {
   const steps: Step[] = [];
