+++ packages/types/src/common.ts
@@ -1,0 +1,3 @@
+export type Action = "click" | "type" | "scroll" | "select";
+
+export type Callback<S = void, T = void> = (data?: S) => T;
